requirements.txt
Flask
python-dotenv
requests
Flask-SQLAlchemy
Jinja2
psycopg2-binary